feat: changed namespace to DatabaseAccessAbstractions, extended corresponding XML comments
