+++ Code/Light.SharedCore/DatabaseAccessAbstractions/IAsyncSession.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Light.SharedCore.DatabaseAccessAbstractions;
+
+/// <summary>
+/// Represents an asynchronous session to a database that manipulates data. This means that
+/// the implementation will either use a single dedicated transaction in the case of ADO.NET or Micro-ORMs
+/// to ensure ACID properties, or the Change Tracking capabilities of a full ORM like Entity Framework Core.
+/// The connection to the database can be terminated by calling
+/// <see cref="IAsyncDisposable.DisposeAsync" /> (or <see cref="IDisposable.Dispose" />), the underlying transaction
+/// will be automatically rolled back if <see cref="SaveChangesAsync"/> was not called beforehand.
+/// Changes can be saved or committed to the database by calling <see cref="SaveChangesAsync" />.
+/// If your session does not manipulate data, consider using the <see cref="IAsyncReadOnlySession" /> interface instead.
+/// </summary>
+/// <remarks>
+/// Conceptually, a session is identical to a "Unit of Work" as defined in "Patterns of
+/// Enterprise Application Architecture" by Martin Fowler et al. It manages the connection
+/// to the database and represents a transaction. Strictly speaking, a Unit of Work also needs
+/// to do Change Tracking which plain ADO.NET and all Micro-ORMs do not support. For this reason,
+/// we chose the term "session" instead of "Unit of Work", also because it is simpler to use in daily life.
+/// </remarks>
+public interface IAsyncSession : IAsyncReadOnlySession
+{
+    /// <summary>
+    /// Commits all changes to the database.
+    /// </summary>
+    /// <param name="cancellationToken">The token to cancel this asynchronous operation (optional).</param>
+    Task SaveChangesAsync(CancellationToken cancellationToken = default);
+}
