+++ Code/Light.SharedCore/DatabaseAccessAbstractions/IAsyncTransaction.cs
@@ -2,11 +2,11 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Light.SharedCore.DataAccessAbstractions;
+namespace Light.SharedCore.DatabaseAccessAbstractions;
 
 /// <summary>
 /// Represents an asynchronous transaction that can be committed. The transaction should always be disposed.
-/// A rollback is performed automatically on <see cref="IAsyncDisposable.DisposeAsync" /> when commit was not
+/// A rollback is performed automatically on <see cref="IAsyncDisposable.DisposeAsync" /> when <see cref="CommitAsync"/> was not
 /// called beforehand.
 /// </summary>
 public interface IAsyncTransaction : IAsyncDisposable, IDisposable
