+++ Code/Light.SharedCore/DatabaseAccessAbstractions/ITransaction.cs
@@ -1,14 +1,18 @@
 ﻿using System;
 
-namespace Light.SharedCore.DataAccessAbstractions;
+namespace Light.SharedCore.DatabaseAccessAbstractions;
 
 /// <summary>
-/// Represents a synchronous transaction that can be committed. The transaction should always be disposed.
-/// A rollback is performed automatically on <see cref="IDisposable.Dispose" /> when commit was not called
-/// beforehand.
+/// <para>
 /// PLEASE REMEMBER: database calls should be performed asynchronously
 /// by default, especially in service apps to avoid blocking threads.
 /// Consider using <see cref="IAsyncTransaction" /> instead.
+/// </para>
+/// <para>
+/// Represents a synchronous transaction that can be committed. The transaction should always be disposed.
+/// A rollback is performed automatically on <see cref="IDisposable.Dispose" /> when commit was not called
+/// beforehand.
+/// </para>
 /// </summary>
 public interface ITransaction : IDisposable
 {
