+++ Code/Light.SharedCore/DatabaseAccessAbstractions/ITransactionalSession.cs
@@ -1,18 +1,23 @@
 ﻿using System;
 
-namespace Light.SharedCore.DataAccessAbstractions;
+namespace Light.SharedCore.DatabaseAccessAbstractions;
 
 /// <summary>
-/// Represents a synchronous session to a database that is able to create individual transactions.
-/// The connection to the database can be terminated by calling <see cref="IDisposable.Dispose" />.
-/// Changes can be committed by first creating a transaction using <see cref="BeginTransaction" />
-/// and later committing this transaction.
+/// <para>
 /// PLEASE REMEMBER: database calls should be performed asynchronously
 /// by default, especially in service apps to avoid blocking threads. Consider using the
 /// <see cref="IAsyncTransactionalSession" /> instead. Furthermore, when your connection to the
 /// database always uses a single transaction, please consider using the <see cref="IAsyncSession" />
-/// or <see cref="ISession" /> interfaces instead. Avoid nesting of several transactions as the interfaces
+/// or <see cref="ISession" /> interfaces instead.
+/// </para>
+/// <para>
+/// Represents a synchronous session to a database that is able to create individual transactions.
+/// The connection to the database can be terminated by calling <see cref="IDisposable.Dispose" />.
+/// Changes can be committed by first creating a transaction using <see cref="BeginTransaction" />
+/// and later committing this transaction.
+/// Avoid nesting of several transactions as the interfaces
 /// in this package do not cover trees of transactions.
+/// </para>
 /// </summary>
 public interface ITransactionalSession : IDisposable
 {
